pyproject.toml
[project]
name = "airbnbclone"
version = "0.1.0"
description = ""
authors = [
    {name = "yusunhyuk"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
